adding updated test_eda script
